refactor: split up utils file and move utils to `util` dir
