+++ src/django_components/autodiscover.py
@@ -7,9 +7,9 @@ from django.apps import apps
 from django.conf import settings
 
 from django_components.app_settings import app_settings
-from django_components.logger import logger
+from django_components.util.logger import logger
 from django_components.template_loader import get_component_dirs
-from django_components.utils import search_dirs
+from django_components.util.utils import search_dirs
 
 
 def autodiscover(
