+++ src/django_components/util/validation.py
@@ -1,50 +1,6 @@
-import functools
-import glob
-import re
 import sys
 import typing
-from pathlib import Path
-from typing import Any, Callable, List, Mapping, Sequence, Tuple, Type, TypeVar, Union, cast, get_type_hints
-
-from django.template.defaultfilters import escape
-from django.utils.autoreload import autoreload_started
-
-from django_components.util.nanoid import generate
-
-
-# Based on nanoid implementation from
-# https://github.com/puyuan/py-nanoid/tree/99e5b478c450f42d713b6111175886dccf16f156/nanoid
-def gen_id() -> str:
-    """Generate a unique ID that can be associated with a Node"""
-    # Alphabet is only alphanumeric. Compared to the default alphabet used by nanoid,
-    # we've omitted `-` and `_`.
-    # With this alphabet, at 6 chars, the chance of collision is 1 in 3.3M.
-    # See https://zelark.github.io/nano-id-cc/
-    return generate(
-        "0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ",
-        size=6,
-    )
-
-
-def find_last_index(lst: List, predicate: Callable[[Any], bool]) -> Any:
-    for r_idx, elem in enumerate(reversed(lst)):
-        if predicate(elem):
-            return len(lst) - 1 - r_idx
-    return -1
-
-
-def is_str_wrapped_in_quotes(s: str) -> bool:
-    return s.startswith(('"', "'")) and s[0] == s[-1] and len(s) >= 2
-
-
-# See https://github.com/EmilStenstrom/django-components/issues/586#issue-2472678136
-def watch_files_for_autoreload(watch_list: Sequence[Union[str, Path]]) -> None:
-    def autoreload_hook(sender: Any, *args: Any, **kwargs: Any) -> None:
-        watch = sender.extra_files.add
-        for file in watch_list:
-            watch(Path(file))
-
-    autoreload_started.connect(autoreload_hook)
+from typing import Any, Mapping, Tuple, get_type_hints
 
 
 # Get all types that users may use from the `typing` module.
@@ -173,95 +129,3 @@ def validate_typed_dict(value: Mapping[str, Any], dict_type: Any, prefix: str, k
         # `Component 'name' got unexpected slot keys 'invalid_key'`
         # `Component 'name' got unexpected data keys 'invalid_key'`
         raise TypeError(f"{prefix} got unexpected {kind} keys {formatted_keys}")
-
-
-TFunc = TypeVar("TFunc", bound=Callable)
-
-
-def lazy_cache(
-    make_cache: Callable[[], Callable[[Callable], Callable]],
-) -> Callable[[TFunc], TFunc]:
-    """
-    Decorator that caches the given function similarly to `functools.lru_cache`.
-    But the cache is instantiated only at first invocation.
-
-    `cache` argument is a function that generates the cache function,
-    e.g. `functools.lru_cache()`.
-    """
-    _cached_fn = None
-
-    def decorator(fn: TFunc) -> TFunc:
-        @functools.wraps(fn)
-        def wrapper(*args: Any, **kwargs: Any) -> Any:
-            # Lazily initialize the cache
-            nonlocal _cached_fn
-            if not _cached_fn:
-                # E.g. `lambda: functools.lru_cache(maxsize=app_settings.TEMPLATE_CACHE_SIZE)`
-                cache = make_cache()
-                _cached_fn = cache(fn)
-
-            return _cached_fn(*args, **kwargs)
-
-        # Allow to access the LRU cache methods
-        # See https://stackoverflow.com/a/37654201/9788634
-        wrapper.cache_info = lambda: _cached_fn.cache_info()  # type: ignore
-        wrapper.cache_clear = lambda: _cached_fn.cache_clear()  # type: ignore
-
-        # And allow to remove the cache instance (mostly for tests)
-        def cache_remove() -> None:
-            nonlocal _cached_fn
-            _cached_fn = None
-
-        wrapper.cache_remove = cache_remove  # type: ignore
-
-        return cast(TFunc, wrapper)
-
-    return decorator
-
-
-def any_regex_match(string: str, patterns: List[re.Pattern]) -> bool:
-    return any(p.search(string) is not None for p in patterns)
-
-
-def no_regex_match(string: str, patterns: List[re.Pattern]) -> bool:
-    return all(p.search(string) is None for p in patterns)
-
-
-def search_dirs(dirs: List[Path], search_glob: str) -> List[Path]:
-    """
-    Search the directories for the given glob pattern. Glob search results are returned
-    as a flattened list.
-    """
-    matched_files: List[Path] = []
-    for directory in dirs:
-        for path in glob.iglob(str(Path(directory) / search_glob), recursive=True):
-            matched_files.append(Path(path))
-
-    return matched_files
-
-
-# See https://stackoverflow.com/a/2020083/9788634
-def get_import_path(cls_or_fn: Type[Any]) -> str:
-    """
-    Get the full import path for a class or a function, e.g. `"path.to.MyClass"`
-    """
-    module = cls_or_fn.__module__
-    if module == "builtins":
-        return cls_or_fn.__qualname__  # avoid outputs like 'builtins.str'
-    return module + "." + cls_or_fn.__qualname__
-
-
-# See https://stackoverflow.com/a/58800331/9788634
-# str.replace(/\\|`|\$/g, '\\$&');
-JS_STRING_LITERAL_SPECIAL_CHARS_REGEX = re.compile(r"\\|`|\$")
-
-
-# See https://stackoverflow.com/a/34064434/9788634
-def escape_js_string_literal(js: str) -> str:
-    escaped_js = escape(js)
-
-    def on_replace_match(match: "re.Match[str]") -> str:
-        return f"\\{match[0]}"
-
-    escaped_js = JS_STRING_LITERAL_SPECIAL_CHARS_REGEX.sub(on_replace_match, escaped_js)
-    return escaped_js
